CoreInterface files and related have been moved to separate directory #403
