+++ sources/Log.cpp
@@ -7,7 +7,7 @@
 #include <QTextStream>
 #include <QUuid>
 
-#include "KvAbort.h"
+#include "core_headers/KvAbort.h"
 
 namespace
 {
