+++ sources/core/Game.cpp
@@ -1,6 +1,6 @@
 #include "Game.h"
 
-#include "KvAbort.h"
+#include "core_headers/KvAbort.h"
 
 #include "Map.h"
