+++ sources/core_headers/CoreInterface.h
