+++ sources/core_headers/FastSerializer.h
@@ -5,7 +5,7 @@
 
 #include <vector>
 
-#include "KvAbort.h"
+#include "core_headers/KvAbort.h"
 
 namespace kv
 {
