+++ sources/core_headers/KvAbort.h
